requirements.txt
tensorflow==2.5.0
streamlit==0.87.0
numpy==1.18.5
pillow==6.2.0